Revert "a"
This reverts commit 3477a49484cf046cbb4c8a1bb5e8bf47f46f5ee0.

+++ backend/src/models/products/products_response.py
@@ -7,7 +7,7 @@ from src.models.key.key import KeyRespond
 class ProductResponse(BaseModel):
     name: Indexed(str,unique=True) # type: ignore 
     description:str
-    price:float  # Changed from int to float to support decimal prices
+    price:int
     active:bool
